@@ -24,9 +24,8 @@ class NetworkAwareWidgetScreen extends StatelessWidget {
           Padding5(),
           ConnectivityWidgetWrapper(
             stacked: false,
-            offlineWidget: RaisedButton(
+            offlineWidget: ElevatedButton(
               onPressed: null,
-              color: Colors.grey,
               child: Padding(
                 padding: const EdgeInsets.all(8.0),
                 child: Row(
@@ -44,7 +43,7 @@ class NetworkAwareWidgetScreen extends StatelessWidget {
                 ),
               ),
             ),
-            child: RaisedButton(
+            child: ElevatedButton(
               onPressed: () {},
               child: Text(
                 "Sign In",
@@ -52,7 +51,6 @@ class NetworkAwareWidgetScreen extends StatelessWidget {
                   color: Colors.white,
                 ),
               ),
-              color: Colors.blue,
             ),
           ),
         ],
